crime report/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pymongo==3.12.0
cryptography==3.4.7
PyJWT==2.1.0
bcrypt==3.2.0
Werkzeug==2.0.1
python-dotenv==0.19.0
requests==2.26.0
